コンポーネント化 pages/items/index
